fix core tests
