+++ packages/codemods/replace-feature-flag-core/__testfixtures__/conditional-expressions.input.ts
@@ -1,4 +1,4 @@
-const isDefaulted = useVariableValue(user, 'simple-case', true);
+const isDefaulted = useFlag(user, 'simple-case', true);
 
 const var1 = isDefaulted ? b : c;
 const var2 = isDefaulted && ( false || true) ? b : c;
