+++ packages/codemods/replace-feature-flag-core/__testfixtures__/object-literal-refactor.input.ts
@@ -1,5 +1,5 @@
 
-const a = useVariableValue(user, 'simple-case', 'string1');
+const a = useFlag(user, 'simple-case', 'string1');
 const b = __CODEMOD_LITERAL__({
     key: "simple-case",
     value: "string1",
