+++ packages/codemods/replace-feature-flag-core/__testfixtures__/references-refactor.input.ts
@@ -1,5 +1,5 @@
-const var2 = useVariableValue(user, 'simple-case', true);
-const var3 = useVariableValue(user, 'simple-case', true) + 'some value';
+const var2 = useFlag(user, 'simple-case', true);
+const var3 = useFlag(user, 'simple-case', true) + 'some value';
 
 const var4 = var2;
 const var5 = var3;
