--- packages/codemods/replace-feature-flag-core/src/providers/index.ts
@@ -1,2 +1,0 @@
-export { DevCycle } from "./dvc.js";
-export { Netlify } from "./netlify.js";
