+++ packages/codemods/replace-feature-flag-core/test/test.ts
@@ -1,10 +1,11 @@
 import assert from "node:assert";
 import { readFile } from "node:fs/promises";
 import { join } from "node:path";
-import { IndentationText, Project } from "ts-morph";
+import { type CallExpression, IndentationText, Project } from "ts-morph";
 import { describe, it } from "vitest";
 import { handleSourceFile } from "../src/index.js";
-import type { Options } from "../src/types.js";
+import type { Options, VariableType, VariableValue } from "../src/types.js";
+import { buildLiteral, getCEExpressionName } from "../src/utils.js";
 
 const transform = (
   projectFiles: Record<string, string>,
@@ -36,43 +37,27 @@ const transform = (
   return transformed;
 };
 
+const fakeProvider = {
+  getMatcher: (keyName: string) => (ce: CallExpression) => {
+    const name = getCEExpressionName(ce);
+
+    if (name !== "useFlag") {
+      return null;
+    }
+
+    return { name };
+  },
+  getReplacer: (
+    key: string,
+    type: VariableType,
+    value: VariableValue,
+    name: string,
+  ) => {
+    return buildLiteral(type, value);
+  },
+};
+
 describe("replace-feature-flag", () => {
-  it("Should replace sdk method call with its return value", async () => {
-    const OUTPUT = await readFile(
-      join(__dirname, "..", "__testfixtures__/simple-string.output.ts"),
-      "utf-8",
-    );
-
-    const projectFiles = {
-      "shared.ts": await readFile(
-        join(__dirname, "..", "__testfixtures__/shared.ts"),
-        "utf-8",
-      ),
-      "simple-string.input.ts": await readFile(
-        join(__dirname, "..", "__testfixtures__/simple-string.input.ts"),
-        "utf-8",
-      ),
-    };
-
-    const stringOptions = {
-      key: "simple-case",
-      type: "string",
-      value: "string",
-      provider: "DevCycle",
-    } as const;
-
-    const transformed = transform(
-      projectFiles,
-      "simple-string.input.ts",
-      stringOptions,
-    );
-
-    assert.deepEqual(
-      transformed?.replace(/\s/gm, ""),
-      OUTPUT?.replace(/\s/gm, ""),
-    );
-  });
-
   it("Should refactor objects", async () => {
     const OUTPUT = await readFile(
       join(
@@ -98,7 +83,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "string",
       value: "string",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -130,7 +115,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -162,7 +147,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -194,7 +179,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -226,7 +211,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -258,7 +243,7 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
@@ -266,8 +251,6 @@ describe("replace-feature-flag", () => {
       "if-statements.input.ts",
       booleanFlagOptions,
     );
-
-    console.log(transformed, "???");
 
     assert.deepEqual(
       transformed?.replace(/\s/gm, ""),
@@ -300,76 +283,12 @@ describe("replace-feature-flag", () => {
       key: "simple-case",
       type: "boolean",
       value: "true",
-      provider: "DevCycle",
+      provider: fakeProvider,
     } as const;
 
     const transformed = transform(
       projectFiles,
       "conditional-expressions.input.ts",
-      booleanFlagOptions,
-    );
-
-    assert.deepEqual(
-      transformed?.replace(/\s/gm, ""),
-      OUTPUT?.replace(/\s/gm, ""),
-    );
-  });
-
-  it("Should refactor javascript", async () => {
-    const OUTPUT = await readFile(
-      join(__dirname, "..", "__testfixtures__/javascript.output.js"),
-      "utf-8",
-    );
-
-    const projectFiles = {
-      "javascript.input.js": await readFile(
-        join(__dirname, "..", "__testfixtures__/javascript.input.js"),
-        "utf-8",
-      ),
-    };
-
-    const booleanFlagOptions = {
-      key: "simple-case",
-      type: "boolean",
-      value: "true",
-      provider: "DevCycle",
-    } as const;
-
-    const transformed = transform(
-      projectFiles,
-      "javascript.input.js",
-      booleanFlagOptions,
-    );
-
-    assert.deepEqual(
-      transformed?.replace(/\s/gm, ""),
-      OUTPUT?.replace(/\s/gm, ""),
-    );
-  });
-
-  it("Should support statsig provider", async () => {
-    const OUTPUT = await readFile(
-      join(__dirname, "..", "__testfixtures__/statsig.output.js"),
-      "utf-8",
-    );
-
-    const projectFiles = {
-      "statsig.input.js": await readFile(
-        join(__dirname, "..", "__testfixtures__/statsig.input.js"),
-        "utf-8",
-      ),
-    };
-
-    const booleanFlagOptions = {
-      key: "the-gate",
-      type: "boolean",
-      value: "true",
-      provider: "Statsig",
-    } as const;
-
-    const transformed = transform(
-      projectFiles,
-      "statsig.input.js",
       booleanFlagOptions,
     );
